+++ bytebank-herdado-conta/src/br/com/bytebank/banco/modelo/ContaCorrente.java
@@ -1,5 +1,5 @@
 //new ContaCorrente()
-package modelo;
+package br.com.bytebank.banco.modelo;
 
 public class ContaCorrente extends Conta implements Tributavel {
