+++ bytebank-herdado-conta/src/br/com/bytebank/banco/modelo/ContaPoupanca.java
@@ -1,4 +1,4 @@
-package modelo;
+package br.com.bytebank.banco.modelo;
 
 public class ContaPoupanca extends Conta {
