+++ bytebank-herdado-conta/src/br/com/bytebank/banco/modelo/SaldoInsuficienteException.java
@@ -1,4 +1,4 @@
-package modelo;
+package br.com.bytebank.banco.modelo;
 
 public class SaldoInsuficienteException extends Exception{
