+++ bytebank-herdado-conta/src/br/com/bytebank/banco/modelo/SeguroDeVida.java
@@ -1,4 +1,4 @@
-package modelo;
+package br.com.bytebank.banco.modelo;
 
 public class SeguroDeVida implements Tributavel {
