+++ bytebank-herdado-conta/src/br/com/bytebank/banco/modelo/Tributavel.java
@@ -1,4 +1,4 @@
-package modelo;
+package br.com.bytebank.banco.modelo;
 
 public interface Tributavel {
