+++ bytebank-herdado-conta/src/br/com/bytebank/banco/teste/TesteContas.java
@@ -1,20 +1,22 @@
-package teste;
+package br.com.bytebank.banco.teste;
+
+import br.com.bytebank.banco.modelo.*;
 
 public class TesteContas {
 
-	public static void main(String[] args) throws modelo.SaldoInsuficienteException{
+	public static void main(String[] args) throws SaldoInsuficienteException{
 		
 		int a = 3;
 		int b = a / 0;
 		
 		//Full Qualified Name FQN
-		modelo.ContaCorrente outra = null;
+		ContaCorrente outra = null;
 		outra.deposita(200.0); 
 					
-		modelo.ContaCorrente cc = new modelo.ContaCorrente(111, 111);
+		ContaCorrente cc = new ContaCorrente(111, 111);
 		cc.deposita(100.0);
 		
-		modelo.ContaPoupanca cp = new modelo.ContaPoupanca(222, 222);
+		ContaPoupanca cp = new ContaPoupanca(222, 222);
 		cp.deposita(200.0);
 		
 		cc.transfere(10.0, cp);
