+++ bytebank-herdado-conta/src/br/com/bytebank/banco/teste/TesteSaca.java
@@ -1,15 +1,17 @@
-package teste;
+package br.com.bytebank.banco.teste;
+
+import br.com.bytebank.banco.modelo.*;
 
 public class TesteSaca {
 
 	public static void main(String[] args) {
-		modelo.Conta conta = new modelo.ContaCorrente(123, 321);
+		Conta conta = new ContaCorrente(123, 321);
 		
 		conta.deposita(200.0);
 		
 		try {
 			conta.saca(210.0);
-		} catch(modelo.SaldoInsuficienteException ex) {
+		} catch(SaldoInsuficienteException ex) {
 			System.out.println("Ex: " + ex.getMessage());
 		}
