+++ bytebank-herdado-conta/src/br/com/bytebank/banco/teste/TesteTributaveis.java
@@ -1,0 +1,21 @@
+package br.com.bytebank.banco.teste;
+
+import br.com.bytebank.banco.modelo.*;
+
+public class TesteTributaveis {
+
+	public static void main(String[] args) {
+		ContaCorrente cc= new ContaCorrente(222, 333);
+		cc.deposita(100.0);
+		
+		SeguroDeVida seguro = new SeguroDeVida();
+		
+		CalculadorDeImposto calc = new CalculadorDeImposto();
+		calc.registra(cc);
+		calc.registra(seguro);
+		
+		System.out.println(calc.getTotalImposto());
+
+	}
+
+}
